refactor(e2e): Extract and reuse integration test utils
Integration test utils were partially the same between v1 and v2 tests, to
simplify future fixes like the recent ones on cleaning up settings and correctly
instantiating dynatrace clients, code from the v2 utils is extracted and reused
as far as possible.

Specific v1 utils are for now left untouched to not blow up the scope of this
refactoring.
