+++ cmd/monaco/integrationtest/assert.go
@@ -1,0 +1,220 @@
+//go:build integration || integration_v1 || cleanup || download_restore || nightly
+
+/*
+ * @license
+ * Copyright 2023 Dynatrace LLC
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package integrationtest
+
+import (
+	"errors"
+	"fmt"
+	"github.com/dynatrace/dynatrace-configuration-as-code/internal/idutils"
+	"github.com/dynatrace/dynatrace-configuration-as-code/internal/log"
+	"github.com/dynatrace/dynatrace-configuration-as-code/internal/testutils"
+	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/api"
+	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/client"
+	config "github.com/dynatrace/dynatrace-configuration-as-code/pkg/config/v2"
+	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/config/v2/coordinate"
+	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/config/v2/parameter"
+	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/deploy"
+	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/manifest"
+	"github.com/dynatrace/dynatrace-configuration-as-code/pkg/project/v2/topologysort"
+	"github.com/spf13/afero"
+	"gotest.tools/assert"
+	"testing"
+	"time"
+)
+
+// AssertAllConfigsAvailability checks all configurations of a given project with given availability
+func AssertAllConfigsAvailability(t *testing.T, fs afero.Fs, manifestPath string, specificProjects []string, specificEnvironment string, available bool) {
+	loadedManifest := LoadManifest(t, fs, manifestPath)
+
+	var specificEnvs []string
+	if specificEnvironment != "" {
+		specificEnvs = append(specificEnvs, specificEnvironment)
+	}
+	environments, err := loadedManifest.Environments.FilterByNames(specificEnvs)
+	if err != nil {
+		t.Fatalf("Failed to filter environments: %v", err)
+	}
+
+	projects := LoadProjects(t, fs, manifestPath, loadedManifest)
+
+	envNames := make([]string, 0, len(environments))
+
+	for _, env := range environments {
+		envNames = append(envNames, env.Name)
+	}
+
+	sortedConfigs, errs := topologysort.GetSortedConfigsForEnvironments(projects, envNames)
+	testutils.FailTestOnAnyError(t, errs, "sorting configurations failed")
+
+	checkString := "exist"
+	if !available {
+		checkString = "do NOT exist"
+	}
+
+	projectsToValidate := map[string]struct{}{}
+	if len(specificProjects) > 0 {
+		log.Info("Asserting configurations from projects: %s %s", specificProjects, checkString)
+		for _, p := range specificProjects {
+			projectsToValidate[p] = struct{}{}
+		}
+	} else {
+		log.Info("Asserting configurations from all projects %s", checkString)
+		for _, p := range projects {
+			projectsToValidate[p.Id] = struct{}{}
+		}
+	}
+
+	for envName, configs := range sortedConfigs {
+
+		env := loadedManifest.Environments[envName]
+
+		c := CreateDynatraceClient(t, env)
+
+		entities := make(map[coordinate.Coordinate]parameter.ResolvedEntity)
+		var parameters []topologysort.ParameterWithName
+
+		for _, theConfig := range configs {
+			coord := theConfig.Coordinate
+
+			if theConfig.Skip {
+				entities[coord] = parameter.ResolvedEntity{
+					EntityName: coord.ConfigId,
+					Coordinate: coord,
+					Properties: parameter.Properties{},
+					Skip:       true,
+				}
+				continue
+			}
+
+			configParameters, errs := topologysort.SortParameters(theConfig.Group, theConfig.Environment, theConfig.Coordinate, theConfig.Parameters)
+			testutils.FailTestOnAnyError(t, errs, "sorting of parameter values failed")
+
+			parameters = append(parameters, configParameters...)
+
+			properties, errs := deploy.ResolveParameterValues(&theConfig, entities, parameters)
+			testutils.FailTestOnAnyError(t, errs, "resolving of parameter values failed")
+
+			properties[config.IdParameter] = "NO REAL ID NEEDED FOR CHECKING AVAILABILITY"
+
+			configName, err := extractConfigName(&theConfig, properties)
+			assert.NilError(t, err)
+
+			entities[coord] = parameter.ResolvedEntity{
+				EntityName: configName,
+				Coordinate: coord,
+				Properties: properties,
+				Skip:       false,
+			}
+
+			apis := api.NewApis()
+			if _, found := projectsToValidate[coord.Project]; found {
+				if theConfig.Type.IsSettings() {
+					assertSetting(t, c, env, available, theConfig)
+				} else if apis.Contains(theConfig.Type.Api) {
+					assertConfig(t, c, apis[theConfig.Type.Api], env, available, theConfig, configName)
+				} else {
+					t.Errorf("Can not assert config of unknown type %q", theConfig.Coordinate.Type)
+				}
+			}
+		}
+	}
+}
+
+func assertConfig(t *testing.T, client client.ConfigClient, theApi api.Api, environment manifest.EnvironmentDefinition, shouldBeAvailable bool, config config.Config, name string) {
+
+	configType := config.Coordinate.Type
+
+	var exists bool
+
+	if config.Skip {
+		exists, _, _ = client.ConfigExistsByName(theApi, name)
+		assert.Check(t, !exists, "Object should NOT be available, but was. environment.Environment: '%s', failed for '%s' (%s)", environment.Name, name, configType)
+		return
+	}
+
+	description := fmt.Sprintf("%s %s on environment %s", configType, name, environment.Name)
+
+	// To deal with delays of configs becoming available try for max 120 polling cycles (4min - at 2sec cycles) for expected state to be reached
+	err := wait(description, 120, func() bool {
+		exists, _, _ = client.ConfigExistsByName(theApi, name)
+		return (shouldBeAvailable && exists) || (!shouldBeAvailable && !exists)
+	})
+	assert.NilError(t, err)
+
+	if shouldBeAvailable {
+		assert.Check(t, exists, "Object should be available, but wasn't. environment.Environment: '%s', failed for '%s' (%s)", environment.Name, name, configType)
+	} else {
+		assert.Check(t, !exists, "Object should NOT be available, but was. environment.Environment: '%s', failed for '%s' (%s)", environment.Name, name, configType)
+	}
+}
+
+func assertSetting(t *testing.T, c client.SettingsClient, environment manifest.EnvironmentDefinition, shouldBeAvailable bool, config config.Config) {
+	expectedExtId := idutils.GenerateExternalID(config.Type.SchemaId, config.Coordinate.ConfigId)
+	objects, err := c.ListSettings(config.Type.SchemaId, client.ListSettingsOptions{DiscardValue: true, Filter: func(o client.DownloadSettingsObject) bool { return o.ExternalId == expectedExtId }})
+	assert.NilError(t, err)
+
+	if len(objects) > 1 {
+		t.Errorf("Expected a specific Settings Object with externalId %q, but %d are present instead.", expectedExtId, len(objects))
+		return
+	}
+
+	exists := len(objects) == 1
+
+	if config.Skip {
+		assert.Check(t, !exists, "Skipped Settings Object should NOT be available but was. environment.Environment: '%s', failed for '%s' (%s)", environment.Name, config.Coordinate, config.Type.SchemaId)
+		return
+	}
+
+	if shouldBeAvailable {
+		assert.Check(t, exists, "Settings Object should be available, but wasn't. environment.Environment: '%s', failed for '%s' (%s)", environment.Name, config.Coordinate, config.Type.SchemaId)
+	} else {
+		assert.Check(t, !exists, "Settings Object should NOT be available, but was. environment.Environment: '%s', failed for '%s' (%s)", environment.Name, config.Coordinate, config.Type.SchemaId)
+	}
+}
+
+func wait(description string, maxPollCount int, condition func() bool) error {
+
+	for i := 0; i <= maxPollCount; i++ {
+
+		if condition() {
+			return nil
+		}
+		time.Sleep(2 * time.Second)
+	}
+
+	log.Error("Error: Waiting for '%s' timed out!", description)
+
+	return errors.New("Waiting for '" + description + "' timed out!")
+}
+
+func extractConfigName(conf *config.Config, properties parameter.Properties) (string, error) {
+	val, found := properties[config.NameParameter]
+
+	if !found {
+		return "", fmt.Errorf("missing `name` for config")
+	}
+
+	name, success := val.(string)
+
+	if !success {
+		return "", fmt.Errorf("`name` in config is not of type string")
+	}
+
+	return name, nil
+}
